requirements.txt
altgraph==0.17.4
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorthief==0.2.1
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
drop==0.4.1
Flask==3.1.0
Flask-Caching==2.3.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.56.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
Panda3D==1.10.15
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pycairo==1.27.0
pygame==2.6.1
PyGObject==3.50.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyngrok==7.2.3
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==3.2.1
pyperclip==1.9.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.2
PyQt6_sip==13.10.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-xlib==0.33
pytz==2025.1
PyYAML==6.0.2
setuptools==76.0.0
six==1.17.0
SQLAlchemy==2.0.39
tk==0.1.0
tkinterdnd2==0.4.3
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
WTForms==3.2.1
